+++ minecraft/src/net/minecraft/client/render/entity/BatEntityRenderer.java
@@ -27,9 +27,9 @@ public class BatEntityRenderer extends MobEntityRenderer<BatEntity, BatEntityMod
 
 	protected void setupTransforms(BatEntity batEntity, MatrixStack matrixStack, float f, float g, float h) {
 		if (batEntity.isRoosting()) {
-			matrixStack.translate(0.0, -0.1F, 0.0);
+			matrixStack.translate(0.0F, -0.1F, 0.0F);
 		} else {
-			matrixStack.translate(0.0, (double)(MathHelper.cos(f * 0.3F) * 0.1F), 0.0);
+			matrixStack.translate(0.0F, MathHelper.cos(f * 0.3F) * 0.1F, 0.0F);
 		}
 
 		super.setupTransforms(batEntity, matrixStack, f, g, h);
