+++ minecraft/src/net/minecraft/client/render/entity/feature/DolphinHeldItemFeatureRenderer.java
@@ -39,9 +39,9 @@ public class DolphinHeldItemFeatureRenderer extends FeatureRenderer<DolphinEntit
 		float n = -1.0F;
 		float o = MathHelper.abs(dolphinEntity.getPitch()) / 60.0F;
 		if (dolphinEntity.getPitch() < 0.0F) {
-			matrixStack.translate(0.0, (double)(1.0F - o * 0.5F), (double)(-1.0F + o * 0.5F));
+			matrixStack.translate(0.0F, 1.0F - o * 0.5F, -1.0F + o * 0.5F);
 		} else {
-			matrixStack.translate(0.0, (double)(1.0F + o * 0.8F), (double)(-1.0F + o * 0.2F));
+			matrixStack.translate(0.0F, 1.0F + o * 0.8F, -1.0F + o * 0.2F);
 		}
 
 		ItemStack itemStack = bl ? dolphinEntity.getMainHandStack() : dolphinEntity.getOffHandStack();
